Fixed error handling in update events in Models
Previously, the erroneous use of `EitherT e (State s) a` prevented
correct handling because when the monad aborts with Left, the state
update has already been recorded. The correct approach is to exchange
the two monads: `StateT s (Either e) a`. Unfortunately (but
reasonably) the Update monad does not have a transformer variant. We
need to define our own. This complicates things, making it necessary
to use Reader for queries as well (for liftQuery). But of course,
correctness is the supreme goal.