Finally FIXed the gradual slowdown bug!
Investigation shows using async in acidQueryHandler to listen for
changes in the database will cause a resource leak i.e. the thread is
never killed, leading to accumulation of items in the TQueue, and
wasting time generating responses. This is fixed by adding another
watchDog thread below. If the queue is not empty for 6 consecutive
times, at intervals of 10 seconds, it is considered that the thread
responsible for sending responses has died due to connection being
closed.

However, another minor problem surfaces, in which a thread that never
writes anything is also not terminated. This is therefore solved by
sending something to the TQueue every 60 seconds regardless of whether
there is such a need.